Refactor ProductMockData

+++ src/helpers/createDataProductMock.js
@@ -1,45 +1,47 @@
-import { useState } from 'react'
-
 import deleteProductMock from '../services/productmocks/deleteProductMock'
 import addProductMock from '../services/productmocks/addProductMock'
 // import getProductMock from '../services/productmocks/getProductMock'
 
-export const createDataProductMock = async ({ url, token, user_id, productMockArray }) => {
-  const [dataProd, setDataProd] = useState([])
+export const createDataProductMock = async ({
+  url,
+  token,
+  user_id,
+  productMockArray,
+  setDataProductMock
+}) => {
   const data = []
 
-  deleteProductMock({ url, token })
+  if (productMockArray.length > 0) {
+    deleteProductMock({ url, token })
 
-  console.log('PRODUCT<MOCK', productMockArray)
+    // console.log('SP productMockArray', productMockArray)
 
-  const dataprd = [...productMockArray]
-  console.log('SP productMockArray', dataprd)
+    for (const item in productMockArray) {
+      const inputAddProduct = {
+        idext: productMockArray[item].extid,
+        amount: productMockArray[item].amount,
+        name: productMockArray[item].shop_list_id,
+        market_id: productMockArray[item].id,
+        barcode: 'XOXO',
+        presentation: 'mock',
+        unit: 'mock',
+        user_id
+      }
 
-  for (const item in dataprd) {
-    const inputAddProduct = {
-      idext: dataprd[item].extid,
-      amount: dataprd[item].amount,
-      name: dataprd[item].shop_list_id,
-      market_id: dataprd[item].id,
-      barcode: 'XOXO',
-      presentation: 'mock',
-      unit: 'mock',
-      user_id
+      addProductMock({
+        url,
+        token,
+        inputAddProduct
+      })
+
+      data.push(inputAddProduct)
     }
-
-    addProductMock({
-      url,
-      token,
-      inputAddProduct
-    })
-
-    // console.log(inputAddProduct.name)
-
-    data.push(inputAddProduct)
+    setDataProductMock(() => data)
+    return data
+  } else {
+    return []
   }
-  setDataProd(() => data)
-  console.log('DATA PRODUCT MOCK', dataProd)
-  return dataProd
+  // console.log('DATA INSIDE createDataProductMock', data)
 }
 
 export default createDataProductMock
